chore: Add .vscode/settings.json file with markdown and typescript configurations
This commit adds a new file ".vscode/settings.json" with configurations for markdown and typescript. The "markdown.copyFiles.destination" setting is added to copy files to a "media" directory with the same name as the file. The "typescript.tsdk" setting is updated to use the "node_modules/typescript/lib" directory. This change improves the development environment for markdown and typescript in the project.