chore: Update ESLint and Prettier configurations
This commit updates the ESLint configuration in the ".eslintrc.js" file to extend the "taro/react" preset and adds the "@typescript-eslint/parser" as the parser. It also updates the Prettier configuration in the ".prettierrc" file to include various formatting rules. These changes improve the code quality and maintainability of the project.